Isolate apiVersion parsing.
